# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/items/{id}": {
      "patch": {
        "operationId": "BoardsController_renameItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rename an item",
        "tags": [
          "boards"
        ]
      },
      "delete": {
        "operationId": "BoardsController_trashItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Move an item to the trash (purged after 30 days unless restored)",
        "tags": [
          "boards"
        ]
      },
      "get": {
        "operationId": "ItemsController_getItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Item detail with cell values, subitems, updates and activity",
        "tags": [
          "items"
        ]
      }
    },
...
    "/v1/boards/{id}/activity": {
      "get": {
        "operationId": "ActivityController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "event",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO date (inclusive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO date (inclusive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Board activity log with filters and keyset pagination",
        "tags": [
          "activity"
        ]
      }
    },
    "/v1/activity/{id}/undo": {
      "post": {
        "operationId": "ActivityController_undo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Undo a change from the last 7 days",
        "tags": [
          "activity"
        ]
      }
    },
...
    "/v1/my-work": {
      "get": {
        "operationId": "ItemsController_myWork",
        "parameters": [
          {
            "name": "includeDone",
            "required": false,
            "in": "query",
            "description": "true to also list completed items",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Items assigned to me, bucketed by due date",
        "tags": [
          "items"
        ]
      }
    },
    "/v1/updates/feed": {
      "get": {
        "operationId": "ItemsController_feed",
        "parameters": [
          {
            "name": "boardId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookmarked",
            "required": false,
            "in": "query",
            "description": "true to show only bookmarked updates",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Company-wide update feed (top-level updates across boards)",
        "tags": [
          "items"
        ]
      }
    },
    "/v1/items/{id}/updates": {
      "post": {
        "operationId": "ItemsController_addUpdate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUpdateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Post an update or a reply (notifies @mentions and the parent author)",
        "tags": [
          "items"
        ]
      }
    },
    "/v1/boards/{id}/updates": {
      "get": {
        "operationId": "ItemsController_boardUpdates",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Board discussion: updates posted on the board itself",
        "tags": [
          "items"
        ]
      },
      "post": {
        "operationId": "ItemsController_addBoardUpdate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUpdateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Post to the board discussion",
        "tags": [
          "items"
        ]
      }
    },
    "/v1/updates/{id}": {
      "patch": {
        "operationId": "ItemsController_editUpdate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit your own update",
        "tags": [
          "items"
        ]
      },
      "delete": {
        "operationId": "ItemsController_deleteUpdate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an update and its replies (author or account admin)",
        "tags": [
          "items"
        ]
      }
    },
    "/v1/updates/{id}/reactions": {
      "post": {
        "operationId": "ItemsController_react",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Toggle an emoji reaction on an update",
        "tags": [
          "items"
        ]
      }
    },
    "/v1/updates/{id}/like": {
      "post": {
        "operationId": "ItemsController_toggleLike",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Toggle your like (a 👍 reaction) on an update",
        "tags": [
          "items"
        ]
      }
    },
    "/v1/updates/{id}/bookmark": {
      "post": {
        "operationId": "ItemsController_toggleBookmark",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Toggle your bookmark on an update",
        "tags": [
          "items"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dayflow API",
    "description": "Work management platform — boards, items, collaboration.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RenameItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "AddUpdateDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "example": "Kicked this off — **draft** ready. @[Alex Smith](user:2c3d…) can you review?",
            "description": "Markdown-lite: **bold**, _italic_, ~~strike~~, `code`, [label](url), \"- \" / \"1. \" lists, @[Name](user:uuid), @[Everyone on this board](board)"
          },
          "parentId": {
            "type": "string",
            "description": "Reply to this top-level update"
          }
        },
        "required": [
          "body"
        ]
      },
      "EditUpdateDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          }
        },
        "required": [
          "body"
        ]
      },
      "ReactDto": {
        "type": "object",
        "properties": {
          "emoji": {
            "type": "string",
            "example": "🎉",
            "description": "One of 👍 ❤️ 🎉 😂 😮 😢 🙏 👀 🔥 ✅"
          }
        },
        "required": [
          "emoji"
        ]
      }
    }
  }
}